+++ tests/runtime/results/wasm.rs
@@ -14,7 +14,7 @@ impl test_exports::Guest for Exports {
         test_imports::string_error(a)
     }
 
-    fn enum_error(a: f64) -> Result<f64, test_exports::E> {
+    fn enum_error(a: f32) -> Result<f32, test_exports::E> {
         match test_imports::enum_error(a) {
             Ok(b) => Ok(b),
             Err(test_imports::E::A) => Err(test_exports::E::A),
@@ -23,14 +23,14 @@ impl test_exports::Guest for Exports {
         }
     }
 
-    fn record_error(a: f64) -> Result<f64, test_exports::E2> {
+    fn record_error(a: f32) -> Result<f32, test_exports::E2> {
         match test_imports::record_error(a) {
             Ok(b) => Ok(b),
             Err(test_imports::E2 { line, column }) => Err(test_exports::E2 { line, column }),
         }
     }
 
-    fn variant_error(a: f64) -> Result<f64, test_exports::E3> {
+    fn variant_error(a: f32) -> Result<f32, test_exports::E3> {
         match test_imports::variant_error(a) {
             Ok(b) => Ok(b),
             Err(test_imports::E3::E1(test_imports::E::A)) => {
